Creep: Added the LoopStream class, the WavePlayer class and the TrackLoader static class
The LoopStream class allows looping a stream between arbitrary points for a specified or unlimited number of repetitions. The range of the playable track in the stream can be changed as well.
The WavePlayer class provides a basic wave stream playback functionality.
The TrackLoader static class provides a convenient way to load a file based on its extension.

Also, VorbisWaveReader has been copied from NAudio.Vorbis source and had its Length and Position properties modified so that they'd work with the LoopStream correctly.