Creep: Added an option to edit basic track information (its name and file path)
The track path can be provided as absolute (mostly for local playlists) and relative (mostly for shareable playlists).
Also, the relative path can be determined based on directory path of the container and the currently known full path.

Also, a bugfix or a few might have made their way in. It might take a while before the application handles all issues and exceptions gracefully, though.

The loop editing is still not supported, but I'm getting there.
Stay tuned!
(haha, get it? Tuned! Because it's a music player! Ahhh, sometimes I amaze myself with my sense of humour...
...but now is not that time u_u')