Creep: Added an option to edit track loop
Now soundtrack range and looping range can be both edited and saved, as long as they pass a basic validation (track start <= loop start < loop end <= track end).
Also, tweaked/fixed a few other things in the process.

On a side note, I just realized that the "Stay tuned!' message from the previous commit didn't serve its purpose at all.
Putting aside dubious popularity of the repository itself, I have added the feature before the commit itself was pushed to origin; so there was literally no-one but me who could see the message in the first place. And I actually planned to wait with pushing to origin in the first place.
It's even more hilarious than the wordplay itself. Not that it's much of an achievement.