Creep: Added an option to save and load playlists
Now it's not necessary to re-add and re-edit tracks every time.
Which is good, because the track edition workflow wasn't designed to be super-quick in the first place.

Also, rearranged things in containers view models, because it turned out synchronised adding/removal/replacement operations are *not* sufficient to keep model and view model synchronised. One needs to *initialise* view model items according to the current model state as well. Details, details...