Creep: Added a playlist reordering through Drag & Drop
It supports reordering a single item as well as a group (the order of items within the group is preserved, no matter the order in which they were selected).
Some elaborate measure has been taken to make it work with selection mechanism and double-clicking, but it works!
I don't really know how to approach keyboard-based reordering. It really should be done (accessibility!), but it'd take some figuring out.

Also, fixed a bug where changing the track range locks causes crashes. The bugs one gets sometimes...