Creep: Added an option to manage the current playlist
In particular, the playlist path can be changed, and it's possible to make the playlist portable or shareable (it may result in completely changing inner items and the playlist's paths).
Portable playlists operate primarily on absolute paths, and are good for moving around own filesystem, but not for sharing between different computers with different directories setup.
Shareable playlists operate solely on relative paths, and are good for sharing between different computers, but must be placed at a specific location in relation to the files it's made for.
If a playlist is neither portable or shareable, it's kinda crappy, since it's cannot really be shared and is stuck in a single location. Please avoid making playlists like that.

Aside from playlist management, several microfeatures, minifixes and nanorefactorings made their way in.