Revision: Changed how Position and Length properties are handled in a limited-repetitions loop
If the track is not set to loop indefinitely, then each individual loop repetition is treated as a separate section. It means the length is increased by the [number of repetitions] * [loop length], and it's possible to move the track to a specific repetition by setting the position within an appropriate range.

Also, now if the playback is stopped, the track will display the position of 0 as well. It's mostly so that when a track stops playing, it isn't shown to be still at the end of the track.