Revision: Changed how limited-repetition loops are displayed in the player
Now they are correctly represented on the player bar, with the middle section being the base loop and the repetition, and the side sections being the intro and outro.

Also, made the player view model update its state and position when the track stops playing on its own.

Also, fixed a bug where the number of total loops wasn't being set when creating a loop stream.
alice is doofus